Added Database Convertion to front end, improved backed stuff.

+++ app/Http/Controllers/Admin/MovieController.php
@@ -57,7 +57,7 @@ class MovieController extends Controller
     public function storeedit(Request $request, $id)
     {
         $movie = Movie::findOrFail($id);
-
+        //dd($request->poster_url);
         $validated = $request->validate([
             'genre_id' => 'required|exists:genres,id',
             'title' => 'required|string|max:100',
@@ -65,10 +65,11 @@ class MovieController extends Controller
             'duration' => 'required|integer',
             'poster_url' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048', // allow nullable for optional re-upload
         ]);
-
-        if ($request->hasFile('poster_url')) {
+        //dd($request->poster_url);
+        if ($request->poster_url) {
             // delete old image if needed
-            if (file_exists(public_path($movie->poster_url))) {
+            //dd($request->hasFile('poster_url'));
+            if ($movie->poster_url && file_exists(public_path($movie->poster_url))) {
                 unlink(public_path($movie->poster_url));
             }
